Auth Updates: Added (1)Reset & (2)Forgot Password -
1. Forgot Password has 3 methods -
(1)Forgot password - Generates a hash as a link to be sent to the email account, which is registered in db.
(2)Verify Forgot Password - Matches & Verifies the email & hash values, from the url, with those in the DB.
(3)Reset Forgot Passord - Resets the account password, if the 'new password' & 'confirm password' matches.

2. Reset Password -
Matches & Verifies the Old password values with DB.
And Resets the account password, if the 'new password' & 'confirm password' matches.

Updated Usertask too.